Move plugin tests to their own file.
